Add web support
This works using `CanvasRenderingContext2D.putImageData`.
The main drawback of this approach is that a canvas can only have one kind of context,
meaning that this will fail if a different kind has already been created,
and will prevent different kinds  being created afterwards.
The first problem could be worked around by adding an implementation for all the different kinds of contexts,
but this is good enough for an MVP.